# api.yaml
openapi: 3.1.0
info:
  title: Project & Task Management API
  description: API for managing companies, users, roles, projects, members, tasks, comments, history, attachments, deadlines, reminders, notifications, and time tracking.
  version: 1.0.0
servers:
  - url: "http://localhost:8080"
security:
  - bearerAuth: []
...
paths:
...
  /tasks:
    get:
      summary: Get all tasks
      tags: [Tasks]
      parameters:
        - in: query
          name: page
          schema:
            type: integer
            minimum: 0
        - in: query
          name: size
          schema:
            type: integer
            minimum: 1
            maximum: 100
        - in: query
          name: status
          schema:
            type: string
        - in: query
          name: project_id
          schema:
            type: string
            format: uuid
        - in: query
          name: assigned_to
          schema:
            type: string
            format: uuid
      responses:
        200:
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Task"
        401:
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"

    post:
      summary: Create new task
      tags: [Tasks]
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TaskCreateRequest"
      responses:
        201:
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Task"
        400:
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        401:
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /tasks/{taskId}:
    get:
      summary: Get task by ID
      tags: [Tasks]
      parameters:
        - $ref: "#/components/parameters/taskId"
      responses:
        200:
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Task"
        401:
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        404:
          description: Not Found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"

    put:
      summary: Update task
      tags: [Tasks]
      parameters:
        - $ref: "#/components/parameters/taskId"
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TaskUpdateRequest"
      responses:
        200:
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Task"
        400:
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        401:
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        403:
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        404:
          description: Not Found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"


    delete:
      summary: Delete task
      tags: [Tasks]
      parameters:
        - $ref: "#/components/parameters/taskId"
      responses:
        204:
          description: Deleted
...
  /tasks/{taskId}/comments:
    get:
      summary: Get comments for task
      tags: [Comments]
      parameters:
        - $ref: "#/components/parameters/taskId"
      responses:
        200:
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Comment"
        401:
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        404:
          description: Not Found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"

    post:
      summary: Add comment
      tags: [Comments]
      parameters:
        - $ref: "#/components/parameters/taskId"
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CommentCreateRequest"
      responses:
        201:
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Comment"
        400:
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        401:
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        404:
          description: Not Found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /tasks/{taskId}/comments/{commentId}:
    delete:
      summary: Delete comment (author or task manager)
      tags: [Comments]
      parameters:
        - $ref: "#/components/parameters/taskId"
        - in: path
          name: commentId
          required: true
          schema:
            type: string
            format: uuid
      responses:
        204:
          description: Deleted
        401:
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        403:
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        404:
          description: Not Found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
...
  /users/{userId}/upcoming:
    get:
      summary: Upcoming deadlines and reminders
      description: Aggregated across own personal projects and all own companies (overdue + next N days, sorted). Powers the dashboard widget.
      tags: [Users]
      parameters:
        - $ref: "#/components/parameters/userId"
        - in: query
          name: days
          schema:
            type: integer
            default: 7
        - in: query
          name: size
          schema:
            type: integer
            default: 20
      responses:
        200:
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Upcoming"
        401:
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        403:
          description: Forbidden (only own items)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
...
  /tasks/{taskId}/attachments:
    get:
      summary: Get attachments for task
      tags: [Attachments]
      parameters:
        - $ref: "#/components/parameters/taskId"
      responses:
        200:
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Attachment"
        401:
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        404:
          description: Not Found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"

    post:
      summary: Upload attachment
      tags: [Attachments]
      parameters:
        - $ref: "#/components/parameters/taskId"
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AttachmentCreateRequest"
      responses:
        201:
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Attachment"
        400:
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        401:
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        404:
          description: Not Found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /tasks/{taskId}/attachments/{attachmentId}:
    put:
      summary: Update attachment
      tags: [Attachments]
      parameters:
        - $ref: "#/components/parameters/taskId"
        - in: path
          name: attachmentId
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AttachmentCreateRequest"
      responses:
        200:
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Attachment"
        400:
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        401:
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        404:
          description: Not Found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"

    delete:
      summary: Delete attachment
      tags: [Attachments]
      parameters:
        - $ref: "#/components/parameters/taskId"
        - in: path
          name: attachmentId
          required: true
          schema:
            type: string
            format: uuid
      responses:
        204:
          description: Deleted
        401:
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        404:
          description: Not Found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
...
components:
  parameters:
    userId:
      name: userId
      in: path
      required: true
      schema:
        type: string
        format: uuid
...
    taskId:
      name: taskId
      in: path
      required: true
      schema:
        type: string
        format: uuid
...
  securitySchemes:
      bearerAuth:
        type: http
        scheme: bearer
        bearerFormat: JWT
  schemas:
...
    Task:
      type: object
      properties:
        id:
          type: string
          format: uuid
        project_id:
          type: string
          format: uuid
        title:
          type: string
        description:
          type: string
        status:
          type: string
          enum: [ OPEN, IN_PROGRESS, DONE, CANCELED]
        priority:
          type: string
          enum: [ LOW, MEDIUM, HIGH, CRITICAL ]
        created_by:
          type: string
          format: uuid
        assigned_to:
          type: string
          format: uuid
        due_date:
          type: string
          format: date-time
        completed_at:
          type: string
          format: date-time
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    TaskCreateRequest:
      type: object
      required: [ project_id, title, status, priority, created_by ]
      properties:
        project_id:
          type: string
          format: uuid
        title:
          type: string
        description:
          type: string
        status:
          type: string
        priority:
          type: string
        created_by:
          type: string
          format: uuid
        assigned_to:
          type: string
          format: uuid
        due_date:
          type: string
          format: date-time
    TaskUpdateRequest:
      type: object
      properties:
        title:
          type: string
        description:
          type: string
        status:
          type: string
        priority:
          type: string
        assigned_to:
          type: string
          format: uuid
        due_date:
          type: string
          format: date-time
        completed_at:
          type: string
          format: date-time
...
    CommentCreateRequest:
      type: object
      required: [ user_id, text ]
      properties:
        user_id:
          type: string
          format: uuid
        text:
          type: string
    Comment:
      type: object
      properties:
        id:
          type: string
          format: uuid
        task_id:
          type: string
          format: uuid
        user_id:
          type: string
          format: uuid
        text:
          type: string
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
...
    Upcoming:
      type: object
      properties:
        deadlines:
          type: array
          items:
            $ref: "#/components/schemas/UpcomingDeadline"
        reminders:
          type: array
          items:
            $ref: "#/components/schemas/UpcomingReminder"
    UpcomingDeadline:
      type: object
      properties:
        id:
          type: string
          format: uuid
        project_id:
          type: string
          format: uuid
        project_name:
          type: string
        title:
          type: string
        deadline:
          type: string
          format: date-time
        overdue:
          type: boolean
    UpcomingReminder:
      type: object
      properties:
        id:
          type: string
          format: uuid
        task_id:
          type: string
          format: uuid
        task_title:
          type: string
        project_id:
          type: string
          format: uuid
        project_name:
          type: string
        remind_at:
          type: string
          format: date-time
        message:
          type: string
        overdue:
          type: boolean
...
    AttachmentCreateRequest:
      type: object
      required: [ file_name, file_url, uploaded_by ]
      properties:
        file_name:
          type: string
        file_url:
          type: string
        uploaded_by:
          type: string
          format: uuid
          description: Ignored on create/update — the uploader is taken from the auth token.
    Attachment:
      type: object
      properties:
        id:
          type: string
          format: uuid
        project_id:
          type: string
          format: uuid
        task_id:
          type: string
          format: uuid
        file_name:
          type: string
        file_url:
          type: string
        uploaded_by:
          type: string
          format: uuid
        uploaded_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
...
    Error:
      type: object
      properties:
        message:
          type: string
        code:
          type: string   